+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### `jsonschema-generator`
+#### Added
+- Convenience methods on `FieldScope`/`MethodScope` for accessing (first level) container item annotations: `getContainerItemAnnotation()` and `getContainerItemAnnotationConsideringFieldAndGetter()`
+
+#### Changed
+- `MethodScope.getAnnotation()` now also considers annotations directly on return type (when no matching annotation was found on the method itself)
+
+### `jsonschema-module-javax-validation`
+#### Changed
+- Consider (first level) container item annotations (e.g. List<@Size(min = 3) String>`)
+
 ## [4.8.1] - 2020-03-31
 ### All
 #### Fixed
@@ -297,6 +309,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.8.1...HEAD
+[4.8.1]: https://github.com/victools/jsonschema-generator/compare/v4.8.0...v4.8.2
 [4.8.0]: https://github.com/victools/jsonschema-generator/compare/v4.7.0...v4.8.0
 [4.7.0]: https://github.com/victools/jsonschema-generator/compare/v4.6.0...v4.7.0
 [4.6.0]: https://github.com/victools/jsonschema-generator/compare/v4.5.0...v4.6.0